Lua: avoid fallback to data dissector for Dissector.call
This makes it possible for call_dissector_work() to drop the protocol
layer for calls to Lua dissectors that repeatedly call Dissector.call to
invoke a Lua dissector which returns 0 to reject data.

While a new Lua API (dissector:call_only) could be added to avoid the
data dissector fallback, it increases complexity of the exposed Lua API.
This patch is a trade-off between:

 - Keep Lua API simple.
 - Show data layer when a dissector is disabled or rejects data.

Right now the latter loses the trade-off in favor of the former, and to
enable Lua dissectors to return 0 to avoid an exception due to too many
Dissector.call calls.

Bug: 15917
Change-Id: If1c03843a6dbe9d373279c7fabd1f3872a7bb424